Add gradual change to pixel opacity
One of the main things to keep in mind is consistency.
Sometimes I have referred to the boxes as boxes, sometimes as pixels.
They are pixels but if I want to use them as such, I should refer to them
as such. Both in my head and in my code. Now it's a bit of a mess with
mentions of both pixel and box. This is to the point where it even annoys
myself

If I am to come back to this project, then this should be the first thing
to fix.

I had some difficulty implementing the gradual change in opacity.
After some console logging of values, it turns out that
this.style.opacity returns a string which then becomes even more string
when trying to += 0.1 to it. Solved by adding a "+" to the reading of
the value from this